requirements.txt
streamlit
openai
faiss-cpu
pandas
openpyxl
google-cloud-bigquery
google-auth
google-auth-oauthlib
pyarrow
db-dtypes
sentence-transformers
plotly
scikit-learn
numpy